- Fixed overflow of Opening Hours for some businesses:   - Updated structure for BusinessDetailsScreen to include TabBar in     AppBar
