+++ lib/features/business/business_details_screen.dart
@@ -1,12 +1,32 @@
 import 'package:flutter/material.dart';
-import 'package:yowl_app/features/business/pages/business_details_page_view.dart';
+import 'package:yowl_app/features/business/pages/business_about_page.dart';
+import 'package:yowl_app/features/business/pages/business_reviews_page.dart';
 import 'package:yowl_app/model/business.dart';
 
-class BusinessDetailsScreen extends StatelessWidget {
+class BusinessDetailsScreen extends StatefulWidget {
   final Business business;
 
   const BusinessDetailsScreen({Key? key, required this.business})
       : super(key: key);
+
+  @override
+  State<BusinessDetailsScreen> createState() => _BusinessDetailsScreenState();
+}
+
+class _BusinessDetailsScreenState extends State<BusinessDetailsScreen>
+    with SingleTickerProviderStateMixin {
+  late TabController _tabController;
+  late List<Tab> _tabs = [];
+
+  @override
+  void initState() {
+    super.initState();
+    _tabs = <Tab>[
+      const Tab(child: Text('About', textAlign: TextAlign.left)),
+      const Tab(child: Text('Reviews', textAlign: TextAlign.left)),
+    ];
+    _tabController = TabController(vsync: this, length: _tabs.length);
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -17,18 +37,48 @@ class BusinessDetailsScreen extends StatelessWidget {
           backgroundColor: Colors.white,
           foregroundColor: Colors.black54,
           title: Text(
-            business.name ?? '',
+            widget.business.name ?? '',
             style: Theme.of(context).textTheme.headlineSmall,
           ),
           centerTitle: true,
+          bottom: _CustomTabBar(
+            color: Colors.white,
+            tabBar: TabBar(
+              padding: const EdgeInsets.symmetric(horizontal: 0.0),
+              controller: _tabController,
+              labelPadding: EdgeInsets.zero,
+              tabs: _tabs,
+              labelColor: Colors.black54,
+            ),
+          ),
         ),
         backgroundColor: Colors.white,
-        body: Column(
-          children: [
-            BusinessDetailsPageView(business: business),
+        body: TabBarView(
+          controller: _tabController,
+          physics: const NeverScrollableScrollPhysics(),
+          children: <Widget>[
+            BusinessAboutPage(business: widget.business),
+            BusinessReviewsPage(business: widget.business),
           ],
         ),
       ),
     );
   }
+}
+
+class _CustomTabBar extends Container implements PreferredSizeWidget {
+  final Color color;
+  final TabBar tabBar;
+
+  _CustomTabBar({Key? key, required this.tabBar, required this.color})
+      : super(key: key);
+
+  @override
+  Size get preferredSize => tabBar.preferredSize;
+
+  @override
+  Widget build(BuildContext context) => Container(
+        color: color,
+        child: tabBar,
+      );
 }
